+++ src/main/kotlin/com/emberjs/actions/EmberGenerateCodeDialog.kt
@@ -1,0 +1,35 @@
+package com.emberjs.actions
+
+import com.emberjs.icons.EmberIconProvider
+import com.emberjs.icons.EmberIcons
+import com.intellij.ide.actions.CreateFileFromTemplateDialog
+import com.intellij.openapi.project.Project
+import javax.swing.Icon
+
+/**
+ * This class is used to expose some of the `protected` properties of the `CreateFileFromTemplateDialog`
+ * base class, that would otherwise not be accessible.
+ */
+class EmberGenerateCodeDialog(project: Project) : CreateFileFromTemplateDialog(project) {
+
+    init {
+        title = "Generate Ember.js Code"
+        setTemplateKindComponentsVisible(true)
+    }
+
+    fun addBlueprint(id: String, name: String = id) {
+        val icon = EmberIconProvider.getIcon(id) ?: EmberIcons.EMPTY_16
+        kindCombo.addItem(name, icon, id)
+    }
+
+    var selectedTemplate: String
+        get() = kindCombo.selectedName
+        set(value) = kindCombo.setSelectedName(value)
+
+    val enteredName: String
+        get() {
+            return nameField.text.trim { it <= ' ' }.apply {
+                nameField.text = this
+            }
+        }
+}